requirements.txt
nselib
pandas
numpy
scipy
requests
psycopg2
yfinance
python-dotenv
